backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
aiofiles==23.2.1
pyttsx3==2.90
asyncio==3.4.3
python-dotenv==1.0.0
aiohttp==3.8.6

# LangChain & LangGraph
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langgraph==0.0.20
langsmith==0.0.69

# Vector Store & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
numpy==1.24.3

# Multi-modal capabilities
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
openai-whisper==20231117
soundfile==0.12.1
pyaudio==0.2.11

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Memory & Storage
redis==5.0.1
sqlalchemy==2.0.23